Loans + preparing for refactor
